requirements.txt
python-dotenv
pyodbc
pandas
scikit-learn
requests
google-genai
